refactor(kfx): isolate runtime warrant adapter
