+++ .buildchain/kfd/support-matrix.json
@@ -534,7 +534,7 @@
         "surfaces": [
           "framework/core/src/libkungfu/src/runtime/kfx/native_registry.cpp",
           "framework/kfx/kungfu-kfx-domain-profile.contract.json",
-          "framework/core/src/python/kungfu/storage/service.py"
+          "framework/core/src/python/kungfu/storage/kfx_service.py"
         ]
       },
       "verification": {
@@ -542,7 +542,7 @@
         "evidenceRoots": [
           {
             "path": "framework/kfx/evidence/kfd-10/runtime-warrant-adopter.json",
-            "sha256": "sha256:46714aef4004b6b40def4881b9d7d4d5e2c8f1b40fe93b9bc140bb911bcfab88"
+            "sha256": "sha256:68a7cb0b50e5f0f4d0aeeac5502d8cfcd1f50bcfd6e1ee5e6770356e84adecde"
           }
         ]
       },
