+++ framework/core/src/python/kungfu/storage/service.py
@@ -8,6 +8,11 @@ from typing import Any
 import kungfu
 
 from kungfu.action_envelope import canonical_json_bytes, payload_hash
+from kungfu.storage.kfx_service import (
+    kfx_registry as _kfx_registry,
+    kfx_runtime_contract as kfx_runtime_contract,
+    validate_kfx_runtime_document as validate_kfx_runtime_document,
+)
 
 PAYLOAD_STATE_PRESENT = "present"
 PAYLOAD_STATES = ("present", "redacted", "absent", "missing")
@@ -869,30 +874,6 @@ def profile_lifecycle(
     )
 
 
-def kfx_runtime_contract(runtime_dir: str | Path = "") -> dict[str, Any]:
-    """Return the versioned Core-owned native KFX contract."""
-
-    return dict(
-        _runtime().run_storage_service_operation(
-            "kfx_runtime", str(runtime_dir), {"action": "contract"}
-        )
-    )
-
-
-def validate_kfx_runtime_document(
-    kind: str, document: dict[str, Any], runtime_dir: str | Path = ""
-) -> dict[str, Any]:
-    """Validate a KFX edge document without reproducing Core policy."""
-
-    return dict(
-        _runtime().run_storage_service_operation(
-            "kfx_runtime",
-            str(runtime_dir),
-            {"action": "validate", "kind": kind, "document": document},
-        )
-    )
-
-
 def kfx_registry(
     action: str,
     request: dict[str, Any],
@@ -900,31 +881,7 @@ def kfx_registry(
 ) -> dict[str, Any]:
     """Project one Core-native KFX registry or lifecycle operation."""
 
-    if action not in {
-        "list",
-        "inspect",
-        "resolve",
-        "plan",
-        "status",
-        "assess",
-        "apply",
-        "history",
-        "authorize-host",
-        "runtime-warrant-issue",
-        "runtime-warrant-heartbeat",
-        "runtime-warrant-revoke",
-        "runtime-warrant-settle",
-        "runtime-warrant-recover",
-        "kfd-10-witness",
-    }:
-        raise ValueError(f"unsupported KFX registry action: {action}")
-    return dict(
-        _runtime().run_storage_service_operation(
-            "kfx_runtime",
-            str(runtime_dir),
-            {"action": action, "request": request},
-        )
-    )
+    return _kfx_registry(action, request, runtime_dir, runtime=_runtime())
 
 
 def fact_contract(runtime_dir: str | Path = "") -> dict[str, Any]:
